Move batch first/last functions in utils (#153)
